+++ src/index.html
@@ -22,15 +22,14 @@
     <!-- Так добавляются фрагменты в главные HTML-файлы страниц. -->
     <!--Header-->
     <include src="./partials/header.html"></include>
-    <!--Search-film-gallery-->
-    <include src="./partials/search-films-gallery.html"></include>
     <!--Loader-->
     <include src="./partials/loader.html"></include>
     <!--Modal-->
     <include src="./partials/modal-auth.html"></include>
     <!--Main-->
     <include src="./partials/main.html"></include>
-
+    <!--Search-film-gallery-->
+    <include src="./partials/search-films-gallery.html"></include>
     <!-- Скрипт идёт последним в разметке. В нём уже подключен главный SASS-файл. -->
     <script src="./index.js"></script>
   </body>
